+++ tipcalculator/TipViewController.m
@@ -43,7 +43,7 @@
 
 - (void)loadUserDefaults {
     NSUserDefaults *defaults = [NSUserDefaults standardUserDefaults];
-    int defaultPercentage = [defaults integerForKey:@"defaultPercentageKey"];
+    float defaultPercentage = [defaults integerForKey:@"defaultPercentageKey"];
     self.tipControl.selectedSegmentIndex = defaultPercentage;
 }
